Link vmsize regex to github data and related code update
